Update slick and fix type error

+++ shared/src/main/scala/org/openmole/connect/shared/Data.scala
@@ -45,7 +45,7 @@ object Data:
     omVersion: String,
     storage: Int,
     memory: Int,
-    cpu: Int,
+    cpu: Double,
     openMOLEMemory: Int,
     lastAccess: Long, created: Long)
